my_agent_utem/requirements.txt
google-adk
google-cloud-aiplatform
google-cloud-discoveryengine
google-cloud-storage
google-cloud-firestore
python-docx
google-auth
fpdf2
pdfplumber

